+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog #
 
+## v1.0.2 ##
+
+* add Heroku and Nginx support and instructions for deployment.
+* add automatic gzip and brotli compression with webpack for serving precompressed static files with express or Nginx.
+* dependencies update
+
 ## v1.0.1 ##
 
 * React-hot-reload on client side and live-reload on server side in combination with Browsersync for testing on external devices.